requirements.txt
absl-py==2.1.0
-e git+https://github.com/flamholz/annotation_free_dioxygen_utilization.git@7fc50336ecb3f6f2e5b31287023cfa70f65bf376#egg=AEROBOT
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
backcall==0.2.0
biopython==1.79
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cycler==0.11.0
datasets==2.13.2
debugpy==1.7.0
decorator==5.1.1
dill==0.3.6
docker-pycreds==0.4.0
entrypoints==0.4
filelock==3.12.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
genslm @ git+https://github.com/ramanathanlab/genslm@e4fbf3b8e641150d708c18e12d551de8ed0cae1c
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
grpcio==1.62.2
h5py==3.7.0
huggingface-hub==0.16.4
idna==3.7
importlib-metadata==6.7.0
ipykernel==6.16.2
ipython==7.34.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.5
lightning-transformers==0.2.1
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
multidict==6.0.5
multiprocess==0.70.14
natsort==8.4.0
nest-asyncio==1.6.0
nltk==3.8.1
numexpr==2.8.6
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
packaging==24.0
pandas==1.3.5
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.0.0
prompt-toolkit==3.0.43
protobuf==3.20.1
psutil==5.9.8
ptyprocess==0.7.0
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.2
pyDeprecate==0.3.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==1.6.5
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
sentencepiece==0.2.0
sentry-sdk==2.2.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
tables==3.7.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.13.1
torchmetrics==0.11.4
tornado==6.2
tqdm==4.66.4
traitlets==5.9.0
transformers @ git+https://github.com/maxzvyagin/transformers@ffd5aba0ad41a1ebd1897a77f6a3782fc2d75e1f
typing_extensions==4.7.1
urllib3==2.0.7
wandb==0.17.0
wcwidth==0.2.13
Werkzeug==2.2.3
wget==3.2
xxhash==3.4.1
yarl==1.9.4
zipp==3.15.0
